requirements/base.txt
Django==1.8.3
django-settings-export==1.0.5
celery==3.1.18
djangorestframework==3.2.0
django-mailgun==0.6.0